refactor: move getDataFromApi into src/utils folder
